walkable bonus room
flow_enter_sub_area pointed the block/coin machinery at the room before the room's own AreaInfo
had been loaded, so it read the coin table through a null level_sub: the room's coins could
never be picked up. the area is loaded first now, and the drop lands on the room's own start
cell. hud_split carried its rest in a uint8_t, which wrapped every score past 25600 - which
1-1's full roster plus its time bonus now clears.

+++ games/mario/src/hud.c
@@ -32,7 +32,9 @@ void hud_split(uint16_t value, uint8_t* out, uint8_t count) {
 
     while (i != 0U) {
         uint16_t rest = value;
-        uint8_t digit = 0;
+        // the carry is the whole rest of the number, not one digit of it: a uint8_t here wrapped
+        // every score past 25600, which 1-1's full sixteen goombas plus its time bonus now clears
+        uint16_t digit = 0;
 
         while (rest >= 10U) {
             rest = (uint16_t)(rest - 10U);
